publicURL: return an explicit component from getPublicUrl
`styled-components` expects elements used in reverse selectors to be
self-contained, executable functions. Wrapping a styled component with
`getPublicUrl()` breaks this, however, as it forces the component to
rely on an injected context.

Componentizing the returned wrapper component from `getPublicUrl()`
results in better error messages if a consumer tries to use the element
in a reverse selector.