SidePanel: add the blocking property
The panel can now be closed using a click on the overlay or the Escape
key.

A new prop, `blocking`, has been added. When set to `true`, the
`onClose` event won’t be invoked, and the close button won’t appear at
all.

It is the responsibility of the developer to implement a closing
mechanism to change the `opened` property.