Move “providers” components to src/providers (#158)
This includes high-order components and render props components.

Also reorganize the exports so that each directory in src/ re-exports its own names.

Each provider is available as a component expecting a render function as its `children`. They also have an `hocWrap` method that acts as an HOC wrapper (this is very useful to wrap styled-components).

The only exception to this is `observe`, which is only an HOC wrapper at the moment and will be migrated later.

In the case of PublicUrl, it is now using the new React Context API (https://reactjs.org/docs/context.html) and require to use `Public.Provider`: `<PublicUrl.Provider url={publicUrl}>`.

Also adds PageRedraw and PageRedrawFromDate pages into the gallery.