Button: introduce the size prop
A new button size needed to be introduced, so this commit introduces a
new prop, `size`, that can be set to `small` or `mini`.

`small` is the equivalent to using the `compact` prop, and should be
used instead.

`mini` is the new size, which is smaller than the `small` size.