Small tweaks
Add two formats: "m" and "s".

Add the showEmpty prop: when set to `true`, the units on the left side
of the timer will be displayed, even if they are equal to zero.

Adjust the time depending on the format, so that the indicated time is
always correct. For example, “1D 3H 10M” becomes “27H 10M” if days are
not displayed.