Button: disabled state + display prop (#432)
The display prop can be used to force the button to behave in certain ways. It replaces iconOnly, which is now deprecated.

display values:

- all: both the icon and the label are displayed.
- icon: only the icon gets displayed.
- label: only the label gets displayed.
- auto (default): like all except in some cases (one for now, see below).

If the Button is inside the secondary slot of Header and its display is set to auto, it will alternate between displaying only the label and only the icon, depending on the current layout.