SidePanel: status, focus, unmount children (#616)
SidePanel is now using Transition, which means that children get properly unmounted when the panel closes. Two new hooks have been added:

## useSidePanel()

This hook should be used inside SidePanel, and returns an object with two entries:

- `status` represents the current state of the panel: opened, closed, opening, closing.
- `readyToFocus` can be used to focus an element as soon as possible inside the panel without affecting the transition.

About the need to have readyToFocus: browser engines tend to force newly focused elements to move inside the viewport, skipping current transitions in the process. readyToFocus detects that the transition is half way, which is generally a safe point to start focusing an element without any transition issue. If you still notice a transition issue, another option is to use status === 'opened'.

This is how readyToFocus can be used:

// MySidePanel.js

function MySidePanel() {
  const inputRef = useRef()
  const { readyToFocus } = useSidePanel()

  useEffect(() => {
    if (readyToFocus && inputRef.current) {
      inputRef.current.focus()
    }
  }, [readyToFocus, inputRef])

  // The TextInput will get focused as soon as possible.
  return <TextInput wide ref={inputRef} />
}

export default props => (
  <SidePanel title="Panel title" {...props}>
    <MySidePanel />
  </SidePanel>
)


And because calling .focus() on a reference is a very common use case, another hook is provided to do it easily: useSidePanelFocusOnReady().

## useSidePanelFocusOnReady()

This hook can be used to focus a ref (usually a TextInput) when readyToFocus is set to true.

It takes a React ref as a parameter, or creates one if not provided. It returns this ref.

// MySidePanel.js

function MySidePanel() {
  const inputRef = useSidePanelFocusOnReady()

  // The TextInput will get focused as soon as possible.
  return <TextInput wide ref={inputRef} />
}

export default props => (
  <SidePanel title="Panel title" {...props}>
    <MySidePanel />
  </SidePanel>
)


## Other changes

- Update README to the new format.
- `title` is now a node rather than a string.
- fix the transition jump near the end (spring precision).