Remove <MainTheme />, replaced by <Main theme />
MainTheme has been removed and should be implemented on the app side if
needed (see devbox/components/current-theme.js or devbox/apps/Theme.js
for examples).

This is because having sub themes in an app is not something people use
at the moment, and being able to set the main theme as a prop makes
things easier to manage for custom themes.

Custom themes are also always extending the base theme corresponding to
their appearance, making the process of extending themes a bit easier.